fix!: handle non integer frame rates adequatly
This fix can be considered as being a breaking change since the results
will be different and that could impact consumers of this library.

The changes were done only for the case where there is no dropped frame.
`allow_inaccurate` was added to `parse_with_fractional_seconds`,
`parse_with_ticks` and `parse` with the default value `value` so those 2
last functions (not `parse`) are backward compatible.

Also fixed the examples shared in the documentation. Even before the
changes of this commit, the 2nd example was wrong. The issues were
```
You are calculating timecodes with different drop flag values (Timecode::WrongDropFlag)
You are calculating timecodes with different framerates (Timecode::WrongFramerate)
```